@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { FormBuilder, FormControl, Validators } from "@angular/forms";
+import { UntypedFormBuilder, UntypedFormControl, Validators } from "@angular/forms";
 import { PasswordChangeModel } from "../models/password-change.model";
 import { ChangePasswordService } from "../services/change-password/change-password.service";
 import { MatSnackBar } from "@angular/material/snack-bar";
@@ -18,9 +18,9 @@ export class ChangePasswordComponent implements OnInit {
   hideNewPassword = true;
   hideConfirmPassword = true;
 
-  oldPassword: FormControl = new FormControl('', Validators.required);
-  newPassword: FormControl = new FormControl('', [Validators.required, Validators.minLength(3)]);
-  confirmPassword: FormControl = new FormControl('', [Validators.required, ExtraValidators.valueMatching(this.newPassword)]);
+  oldPassword: UntypedFormControl = new UntypedFormControl('', Validators.required);
+  newPassword: UntypedFormControl = new UntypedFormControl('', [Validators.required, Validators.minLength(3)]);
+  confirmPassword: UntypedFormControl = new UntypedFormControl('', [Validators.required, ExtraValidators.valueMatching(this.newPassword)]);
 
   changePasswordForm = this.formBuilder.group({
     oldPassword: this.oldPassword.value,
@@ -30,7 +30,7 @@ export class ChangePasswordComponent implements OnInit {
 
 
   constructor(
-    private formBuilder: FormBuilder,
+    private formBuilder: UntypedFormBuilder,
     private changePasswordService: ChangePasswordService,
     private snackBar: MatSnackBar,
     public readonly errorMessageService: FormErrorMessageService,
